backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
numpy==1.26.3
python-socketio
python-multipart
GROQ
opencv-python-headless
mediapipe
numpy